% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dplyr.R
\name{addMeta}
\alias{addMeta}
\title{Add meta data to a data frame}
\usage{
addMeta(df, meta)
}
\arguments{
\item{df}{A data frame}

\item{meta}{A list of attributes}
}
\value{

}
\description{
Add meta data to a data frame
}
\examples{
addMeta(df, list("source"="WDI", "units"="millions"))
}
