% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drake.R
\name{clip_planWrapper}
\alias{clip_planWrapper}
\title{Send a drake plan wrapper to clipboard to paste}
\usage{
clip_planWrapper()
}
\value{

}
\description{
Send a drake plan wrapper to clipboard to paste
}
\examples{
none
}
