% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drake.R
\name{create_planRmd}
\alias{create_planRmd}
\title{Create source_plan function source file under planPath}
\usage{
create_planRmd(planname, title = NULL, root = NULL)
}
\arguments{
\item{planname}{A character.}

\item{title}{A character. (if null, default=planname)}

\item{root}{A chracter. (if null, default=getwd())}
}
\value{
A function that source an R script under planPath


}
\description{
Create source_plan function source file under planPath
}
\examples{
# # NOT RUN. produce error
# create_source_plan()

planPath ="/Users/martin/Github/course-HS-society-and-citizen/plans"
Create Rmd template that can generate drake plan scripts

none
}
