% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drake.R
\name{create_source_plan}
\alias{create_source_plan}
\title{Create source_plan function source file under planPath}
\usage{
create_source_plan()
}
\value{
A function that source an R script under planPath
}
\description{
Create source_plan function source file under planPath
}
\examples{
# # NOT RUN. produce error
# create_source_plan()

planPath ="/Users/martin/Github/course-HS-society-and-citizen/plans"
source_plan <- create_source_plan()

}
