% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplots.R
\name{gen_df_geom_rect}
\alias{gen_df_geom_rect}
\title{產生geom_rect所需xmin, xmax, ymin, ymax dataframe}
\usage{
gen_df_geom_rect(df_x)
}
\arguments{
\item{df_x}{A data frame with xmin and xmax columns}
}
\value{
A data frame with xmin and xmax inherited from df_x, and ymin=-Inf,
ymax=Inf. If xmin/xmax is not date class, will be converted accordingly.
}
\description{
產生geom_rect所需xmin, xmax, ymin, ymax dataframe
}
\examples{
none
}
