% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplotHack.R
\name{get_allDiscrepancyExpressions}
\alias{get_allDiscrepancyExpressions}
\title{Get all discrepancy extraction expressions}
\usage{
get_allDiscrepancyExpressions(list_discrepancy, referenceObject)
}
\arguments{
\item{list_discrepancy}{A list of return value from get_discrepancyExtractionIndex}

\item{referenceObject}{A ggplot object}
}
\value{
A list of discrepancy component value extraction expressions
}
\description{
Get all discrepancy extraction expressions
}
\examples{
none.
}
