% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplotHack.R
\name{get_allDiscrepancyExpressionsInText}
\alias{get_allDiscrepancyExpressionsInText}
\title{Get all discrepancy extractions expression in text}
\usage{
get_allDiscrepancyExpressionsInText(list_discrepancy, referenceObject)
}
\arguments{
\item{list_discrepancy}{A list of return value from get_discrepancyExtractionIndex}

\item{referenceObject}{A ggplot object}
}
\value{
A character vector of discrepancy component value extraction expression in text
}
\description{
Get all discrepancy extractions expression in text
}
\examples{
none.
}
