% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplotHack.R
\name{get_discrepancyExtractionIndex}
\alias{get_discrepancyExtractionIndex}
\title{Get discrepancy extraction indices from all.equal comparison between two ggplot objects}
\usage{
get_discrepancyExtractionIndex(object1, object2)
}
\arguments{
\item{object1}{A ggplot object.}

\item{object2}{A ggplot object.}
}
\value{
A list of indices
}
\description{
Get discrepancy extraction indices from all.equal comparison between two ggplot objects
}
\examples{
none.
}
