% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplots.R
\name{get_rankNumber}
\alias{get_rankNumber}
\title{Convert series into 1, 2, 3 rank numbers. 1 means the highest.}
\usage{
get_rankNumber(series)
}
\arguments{
\item{series}{A numeric vector}
}
\value{
An integer vector, showing the rank, 1 means the larges/highest.
}
\description{
Convert series into 1, 2, 3 rank numbers. 1 means the highest.
}
\examples{
none
}
