% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{get_rdb}
\alias{get_rdb}
\title{Obtain dataset from dbnomics}
\usage{
get_rdb(rdb0, queryString)
}
\arguments{
\item{rdb0}{A list. Returned value from rdbnomics::rdb(api_link=....)}

\item{queryString}{A character.}
}
\value{
data base from the rdb0 dataset that fits queryString query.
}
\description{
Obtain dataset from dbnomics
}
\examples{
none
}
