% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SOAR.R
\name{\%=\%}
\alias{\%=\%}
\title{Enchanced `=` to store object in R.cache}
\usage{
obj \%=\% expr
}
\arguments{
\item{obj}{A symbol or name character represents the object to be created.}

\item{expr}{An expression command that creates the value for the object.}
}
\value{
Null.
}
\description{
Enchanced `=` to store object in R.cache
}
\examples{
none
}
