% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initiate.R
\name{init_courseSetup}
\alias{init_courseSetup}
\title{首次啟動課程環境設定}
\usage{
init_courseSetup(packages_init = econDV::packages_init)
}
\arguments{
\item{packages_init}{A character vector, specifying what packages the course needs}
}
\value{

}
\description{
首次啟動課程環境設定
}
\examples{
none
}
