% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drake.R
\name{list_makes}
\alias{list_makes}
\title{List makes (future) in your global environment}
\usage{
list_makes()
}
\value{
character string of promises of make/mk_plan...
}
\description{
List makes (future) in your global environment
}
\examples{
none
}
