% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drake.R
\name{list_plans}
\alias{list_plans}
\title{List available plans in your global environment}
\usage{
list_plans()
}
\value{
A character string of drake plans
}
\description{
List available plans in your global environment
}
\examples{
none
}
