% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Map.R
\name{osm_geom_rename}
\alias{osm_geom_rename}
\title{A fixture to OSM simple feature date name error problem}
\usage{
osm_geom_rename(sf_object)
}
\arguments{
\item{sf_object}{An sf object}
}
\value{

}
\description{
A fixture to OSM simple feature date name error problem
}
\examples{
none
}
