% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drake.R
\name{purl_drakePlan}
\alias{purl_drakePlan}
\title{Purl Rmd to a drake plan R script}
\usage{
purl_drakePlan(filename, plan_name)
}
\arguments{
\item{filename}{A character defines the file path name of Rmd}

\item{plan_name}{A character defines the R script name to be plan_name.R}
}
\value{
invisible. A character of all R script lines
}
\description{
All R chunks with chunk names without drake=F will be purled to
a Drake plan R script.
}
\examples{
none
}
