% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drake.R
\name{purlActiveRmd_thenPlanMake}
\alias{purlActiveRmd_thenPlanMake}
\title{Purl active Rmd into a drake plan R script, named "plan_{activeRmdName}.R", then make the plan.}
\usage{
purlActiveRmd_thenPlanMake()
}
\value{

}
\description{
Purl active Rmd into a drake plan R script, named "plan_{activeRmdName}.R", then make the plan.
}
\examples{
purlActiveRmd_thenPlanMake()
}
