% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{read2clip}
\alias{read2clip}
\title{Read filelines to clipboard}
\usage{
read2clip(filepath)
}
\arguments{
\item{filepath}{A character. File to readLines from}
}
\value{
content in clipboard for pasting
}
\description{
Read filelines to clipboard
}
\examples{
none
}
