% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{save2_functional}
\alias{save2_functional}
\title{Create save functions with designated saving directory}
\usage{
save2_functional(savepath)
}
\arguments{
\item{savepath}{A character, represents saving directory}
}
\value{
none
}
\description{
Create save functions with designated saving directory
}
\examples{
save2img <- save2_functional("img");
}
