% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmarkdown.R
\name{setup_chinese}
\alias{setup_chinese}
\title{Set up chinese font environment}
\usage{
setup_chinese(need2Knit = F)
}
\arguments{
\item{need2Knit}{A logical, default=T includes fig.showtext=T and fig.retina=1}
}
\value{

}
\description{
Set up chinese font environment
}
\examples{
none
}
