% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplotHack.R
\name{show_values}
\alias{show_values}
\title{Show discrepancy extraction expression on screen}
\usage{
show_values(discrepancyExpression, needEval = T)
}
\arguments{
\item{discrepancyExpression}{A list of return value from get_allDiscrepancyExpressions/inText}

\item{needEval}{logical, default=T, also show component values on screen}
}
\value{

}
\description{
Show discrepancy extraction expression on screen
}
\examples{
none
}
