% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dplyr.R
\name{str_detect_all}
\alias{str_detect_all}
\title{Predicate if detect all possible regex patterns}
\usage{
str_detect_all(series, regex)
}
\arguments{
\item{series}{A character vector}

\item{regex}{A character of regex pattern with "pattern1|pattern2|...|patternK"}
}
\value{
A logical vector. TRUE if fits ALL patterns.
}
\description{
Predicate if detect all possible regex patterns
}
\examples{
none
}
