% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/svg.R
\name{svgOpen}
\alias{svgOpen}
\title{Launch Boxy SVG and open filename inside}
\usage{
svgOpen(filename)
}
\arguments{
\item{filename}{}
}
\value{

}
\description{
Launch Boxy SVG and open filename inside
}
\examples{
none
}
