% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/themes.R
\name{theme_timeSeries}
\alias{theme_timeSeries}
\title{Create a theme for time series that fit the format of the Economist}
\usage{
theme_timeSeries()
}
\value{

}
\description{
Create a theme for time series that fit the format of the Economist
}
\examples{
none
}
